+++ src/auth_config.json
@@ -1,0 +1,4 @@
+{
+  "domain": "envisage-dt.us.auth0.com",
+  "clientId": "m3oHgSAh2FFnzwpICzzf66hFeVGj9KAA"
+}
