+++ src/components/Home/Calendar.css
@@ -2,7 +2,6 @@
   column-gap: 1vw;
   padding: 0;
   border: 0;
-  z-index: 2;
   display: flex;
   align-items: center;
   justify-content: center;
@@ -15,6 +14,11 @@
 
 .calendar .react-datepicker-wrapper {
   width: fit-content;
+  z-index: 2;
+}
+
+.calendar .react-datepicker-popper {
+  z-index: 2;
 }
 
 @media only screen and (min-width: 320px) {
