+++ src/components/Home/DetailsTable.css
@@ -43,26 +43,30 @@
 .tableDiv {
   display: block;
   overflow: auto;
-  height: 75vh;
-  width: 90vw;
+  height: 100%;
+  width: 100%;
   top: 0;
-  z-index: 2;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
+  z-index: 3;
+  position: fixed;
 }
 
 .closeTable {
-  position: fixed;
-  right: 1vw;
+  position: absolute;
+  right: 12.5vw;
+  top: 5vh;
   z-index: 5;
+  border: 1px solid black;
+  background: aliceblue;
 }
 
 .tableInnerDiv {
-  width: 100%;
-  height: 100%;
+  width: 75vw;
+  height: 90vh;
   overflow: auto;
+  top: 50%;
+  left: 50%;
+  position: absolute;
+  transform: translate(-50%, -50%);
 }
 
 table {
