+++ src/components/Home/OutlineCard.css
