+++ src/components/Login/login.css
@@ -1,0 +1,14 @@
+.login {
+  background-color: #35455f;
+  width: 100vw;
+  height: 100vh;
+  color: white;
+}
+
+.loginLoader {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100vw;
+  height: 100vh;
+}
